requirements.txt
keras==2.15.0
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
pandas==2.1.4
Pillow==10.1.0
requests==2.31.0
scikit-learn==1.3.2
tensorflow==2.15.0


